make file naming consistent
