+++ jest.setup.ts
@@ -1,4 +1,4 @@
-import { getDbClient } from "@/app/backend/dbClient";
+import { internal_getDbClient } from "@/app/backend/db/config";
 import "@testing-library/jest-dom";
 import { toHaveNoViolations } from "jest-axe";
 import * as matchers from "jest-extended";
@@ -10,7 +10,7 @@ expect.extend(matchers);
 if (process.env.TEST_TYPE === "integration") {
   let dbClient: Pool | null = null;
   beforeAll(() => {
-    dbClient = getDbClient();
+    dbClient = internal_getDbClient();
   });
 
   afterAll(async () => {
