+++ src/app/(pages)/auditLogs/components/auditLogDrawer.tsx
@@ -2,9 +2,9 @@
 
 import React from "react";
 import Drawer from "@/app/ui/designSystem/drawer/Drawer";
-import { LogEntry } from "@/app/backend/dbServices/audit-logs";
 import styles from "../auditLogs.module.scss";
 import { auditLogActionTypeMap, auditLogUserMap } from "./auditLogMaps";
+import { LogEntry } from "@/app/backend/audit-logs";
 
 type AuditLogDrawerProps = {
   isOpen: boolean;
