+++ src/app/(pages)/query/components/selectQuery/SelectSavedQuery.tsx
@@ -6,18 +6,18 @@ import { RETURN_LABEL } from "@/app/(pages)/query/components/stepIndicator/StepI
 import TitleBox from "../stepIndicator/TitleBox";
 import LoadingView from "../../../../ui/designSystem/LoadingView";
 import { showToastConfirmation } from "../../../../ui/designSystem/toast/Toast";
-import { getFhirServerNames } from "@/app/backend/dbServices/fhir-servers";
 import { CustomUserQuery } from "@/app/models/entities/query";
 import {
   getCustomQueries,
   getQueriesForUser,
-} from "@/app/backend/query-building";
+} from "@/app/backend/query-building/service";
 import { User, UserRole } from "@/app/models/entities/users";
 import { getRole } from "@/app/(pages)/userManagement/utils";
 import { getUserByUsername } from "@/app/backend/user-management";
 import { useSession } from "next-auth/react";
 import { isAuthDisabledClientCheck } from "@/app/utils/auth";
 import { DataContext } from "@/app/shared/DataProvider";
+import { getFhirServerNames } from "@/app/backend/fhir-servers";
 
 type SelectSavedQueryProps = {
   selectedQuery: CustomUserQuery;
