+++ src/app/(pages)/queryBuilding/QueryBuilding.test.tsx
@@ -1,9 +1,9 @@
 import QueryBuilding from "./page";
 import { render, screen, waitFor } from "@testing-library/react";
-import { getQueryList } from "@/app/backend/query-building";
+import { getQueryList } from "@/app/backend/query-building/service";
 import { RootProviderMock } from "@/app/tests/unit/setup";
 
-jest.mock("@/app/backend/query-building", () => ({
+jest.mock("@/app/backend/dbServices/queryBuilding/service", () => ({
   getQueryList: jest.fn(),
 }));
