+++ src/app/(pages)/userManagement/components/teamQueryEditSection/TeamQueryEditSection.test.tsx
@@ -21,7 +21,7 @@ jest.mock(
     ({ children }: React.PropsWithChildren) => <div>{children}</div>,
 );
 
-jest.mock("@/app/backend/query-building", () => ({
+jest.mock("@/app/backend/dbServices/queryBuilding/service", () => ({
   getCustomQueries: jest.fn(),
 }));
