+++ src/app/tests/integration/audit-log/general.test.ts
@@ -6,7 +6,7 @@ import { patientDiscoveryQuery } from "@/app/backend/query-execution";
 import { PatientDiscoveryRequest } from "@/app/models/entities/query";
 import { hyperUnluckyPatient } from "@/app/shared/constants";
 import * as DecoratorUtils from "@/app/backend/auditLogs/lib";
-import { getDbClient } from "@/app/backend/dbClient";
+import { internal_getDbClient } from "@/app/backend/db/config";
 import { suppressConsoleLogs } from "../fixtures";
 
 jest.mock("@/app/backend/auditLogs/lib", () => {
@@ -16,7 +16,7 @@ jest.mock("@/app/backend/auditLogs/lib", () => {
   };
 });
 
-const dbClient = getDbClient();
+const dbClient = internal_getDbClient();
 describe("checks for generic audit logs", () => {
   beforeAll(() => {
     suppressConsoleLogs();
