+++ src/app/tests/integration/usergroup-management.test.ts
@@ -7,12 +7,12 @@ import {
   removeQueriesFromGroup,
 } from "@/app/backend/usergroup-management";
 import { getAllUsersWithSingleGroupStatus } from "@/app/backend/user-management";
-import { getDbClient } from "@/app/backend/dbClient";
+import { internal_getDbClient } from "@/app/backend/db/config";
 import { User } from "@/app/models/entities/users";
 import { suppressConsoleLogs } from "./fixtures";
 import { QueryDataColumn } from "@/app/(pages)/queryBuilding/utils";
 
-const dbClient = getDbClient();
+const dbClient = internal_getDbClient();
 
 jest.mock("@/app/utils/auth", () => ({
   superAdminAccessCheck: jest.fn(() => Promise.resolve(true)),
@@ -166,8 +166,8 @@ describe("User Group and Query Membership Tests", () => {
    */
   test("should remove multiple users from a group", async () => {
     const users: User[] = await getAllUsersWithSingleGroupStatus(TEST_GROUP_ID);
-    const members = users.filter((user) =>
-      user.userGroupMemberships?.some((m) => m.isMember),
+    const members = users.filter(
+      (user) => user.userGroupMemberships?.some((m) => m.isMember),
     );
     expect(members.length).toBe(3);
 
@@ -203,8 +203,8 @@ describe("User Group and Query Membership Tests", () => {
     expect(result.items.length).toBe(1);
 
     const users: User[] = await getAllUsersWithSingleGroupStatus(TEST_GROUP_ID);
-    const members = users.filter((user) =>
-      user.userGroupMemberships?.some((m) => m.isMember),
+    const members = users.filter(
+      (user) => user.userGroupMemberships?.some((m) => m.isMember),
     );
 
     expect(members.length).toBe(0);
